v0-alpha-kotlin: Kotlin migration fixes, Chaquopy 15.0.1, and 7 critical bug fixes
- Fixed: null!! crash on Android 16 (lateinit var for INSTANCE/DATABASE)
- Fixed: Web proxy port hardcoded to 7125 (dynamic moonraker.conf port)
- Fixed: beeper.cpp memory corruption (sizeof double -> float)
- Fixed: Python init infinite retry loop (max 10 attempts + logging)
- Fixed: Silent exception swallowing (Log.w added to 3 catch blocks)
- Fixed: Thread priority -20 (changed to -4, non-real-time)
- Fixed: beam_camera.py wrong log message (flashlight -> focus)
- Fixed: OkHttp 3.12.6 -> 3.14.9
- Fixed: buildBadges ArrayIndexOutOfBoundsException (null array filter)
- Upgraded: Chaquopy 14.0.2 -> 15.0.1
- Patched: streaming-form-data C extension -> pure-Python wheel
- Removed: 74 unsafe !! operators across 22 files

+++ app/src/main/java/ru/ytkab0bp/beamklipper/view/BoostySubsView.kt
@@ -22,7 +22,7 @@ class BoostySubsView(context: Context) : View(context) {
     private val rect = Rect()
 
     init {
-        paint.textSize = ViewUtils.dp(20)
+        paint.textSize = ViewUtils.dp(20).toFloat()
         paint.typeface = ViewUtils.getTypeface(ViewUtils.ROBOTO_MEDIUM)
         updateColors()
     }
